Create Decision Documentation
Several earlier classifications that I discuss are not relevant anymore because the sentiment has since been changed on the spreadsheet after a review. I decided to still include these particular things in the document so that I have correct numbering on the sentences and space to go back and correct when I do my final review. So for now, I know sentences 60-95 are good for review and comments.